+++ example/lib/gallery/sections/sections.dart
@@ -17,4 +17,5 @@ export 'split_panel.dart';
 export 'tab_view.dart';
 export 'text_fields.dart';
 export 'texts.dart';
+export 'tooltips.dart';
 export 'windows.dart';